% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smmry_api.R
\name{smmry_api}
\alias{smmry_api}
\title{test}
\usage{
smmry_api(text = NULL, url = NULL, length = NULL, keywords = NULL,
  quote_avoid = FALSE, breaks = FALSE)
}
\arguments{
\item{text}{test}

\item{url}{test}

\item{length}{test}

\item{keywords}{test}

\item{quote_avoid}{test}

\item{breaks}{test}
}
\description{
test
}
\examples{
lorem_ipsum <- "Lorem ipsum dolor sit amet, consetetur 
         sadipscing elitr, sed diam nonumy eirmod 
         tempor invidunt ut labore et dolore magna 
         aliquyam erat, sed diam voluptua. At vero 
         eos et accusam et justo duo dolores et ea 
         rebum. Stet clita kasd gubergren, no sea 
         takimata sanctus est Lorem ipsum dolor sit 
         amet. Lorem ipsum dolor sit amet, consetetur 
         sadipscing elitr, sed diam nonumy eirmod 
         tempor invidunt ut labore et dolore magna 
         aliquyam erat, sed diam voluptua. At vero 
         eos et accusam et justo duo dolores et ea 
         rebum. Stet clita kasd gubergren, no sea 
         takimata sanctus est Lorem ipsum dolor sit 
         amet."

smmry_api(text = lorem_ipsum)

}
